requirements.txt
langgraph>=0.2.50
gitpython>=3.1.40
click>=8.1.7
pyyaml>=6.0.1
requests>=2.31.0
